NtupleAnalysis DataFormat: reverted fNUP,fNPU,fSimulatedNPU back to short

+++ NtupleAnalysis/src/DataFormat/interface/EventID.h
@@ -24,7 +24,7 @@ public:
   unsigned int       lumi()  const { return fLumi->value(); }
   unsigned int       run()   const { return fRun->value(); }
   /// Number of partons generated
-  int NUP() const { return fNUP->value(); }
+  short NUP() const { return fNUP->value(); }
   /// Trigger prescale
   float trgPrescale() const { return fPrescale->value(); }
   
@@ -32,7 +32,7 @@ private:
   const Branch<unsigned long long> *fEvent;
   const Branch<unsigned int> *fLumi;
   const Branch<unsigned int> *fRun;
-  const Branch<int> *fNUP;
+  const Branch<short> *fNUP;
   const Branch<float> *fPrescale;
 };
